Major Refactor - Part 3
