+++ refactor_video_call/components/index.js
@@ -1,2 +1,2 @@
 export { default as ChatBox } from './ChatBox';
-export { default as NavBar } from './Navbar';
+export { default as NavBar } from './NavBar';
